fix: correct hbool_t type from c_uint (4 bytes) to u8 (1 byte)
HDF5's hbool_t is typedef'd as bool (_Bool), which is 1 byte on all
modern systems with <stdbool.h>. Our definition was c_uint (4 bytes),
causing struct layout mismatches in H5AC_cache_config_t and other
structs containing hbool_t fields.

This was the root cause of the SIGSEGV in test_plist on x86_64 Linux:
H5Pget_mdc_config wrote into H5AC_cache_config_t using 1-byte bool
offsets, but Rust read fields at 4-byte uint offsets, causing
decr_mode to contain an invalid enum discriminant.

Also removes the temporary GDB debugging CI step and continue-on-error.

+++ hdf5/src/sys/runtime.rs
@@ -23,8 +23,8 @@ pub use libc::{
 pub type hid_t = i64;
 /// HDF5 error return type
 pub type herr_t = c_int;
-/// HDF5 boolean type
-pub type hbool_t = c_uint;
+/// HDF5 boolean type (`_Bool` in C, 1 byte on all modern systems with `<stdbool.h>`)
+pub type hbool_t = u8;
 /// HDF5 size type (unsigned)
 pub type hsize_t = c_ulong;
 /// HDF5 signed size type
